+++ assets/css/index.ts
@@ -1,6 +1,1 @@
-import "./primercss.scss";
-import "./actionlist/index.scss";
-import "./prompt.css";
-import "./primercss-overrides.css";
-import "./toggle-switch.css";
-import "./util.css";
+import "./index.scss";