+++ test/assets_test.exs
@@ -5,11 +5,13 @@ defmodule PrimerLiveAssetsTest do
   test "priv/static files" do
     filenames = [
       "index.d.ts",
-      "primer-live.cjs.js",
+      "primer-live-scoped.css",
+      "primer-live-scoped.min.css",
       "primer-live.cjs.js.map",
+      "primer-live.cjs.js",
       "primer-live.css",
-      "primer-live.esm.js",
       "primer-live.esm.js.map",
+      "primer-live.esm.js",
       "primer-live.js",
       "primer-live.min.css",
       "primer-live.min.js"