+++ lib/theme/theme.ex
@@ -1,21 +1,35 @@
 defmodule PrimerLive.Theme do
   @moduledoc """
-  Primer CSS contains styles for light/dark color modes and themes, with support for color blindness.
+  PrimerLive contains styles for light and dark color modes and themes, with support for color blindness.
 
-  PrimerLive provides components and functions to work with themes:
+  PrimerLive provides these theme components and functions:
   - [`theme/1`](`PrimerLive.Component.theme/1`) - wrapper to set the theme on child elements
   - [`theme_menu_options/1`](`PrimerLive.Component.theme_menu_options/1`) - contents for a theme menu
   - `html_attributes/2` - HTML attributes to set a theme on a component or element directly
 
+  ## Theme component
+
+  The `theme/1` component sets the required HTML attributes.
+
+  Default settings (light theme):
+
+  ```
+  <.theme>
+    Content
+  </.theme>
+  ```
+
+  To hardcode the theme to dark mode:
+
+  ```
+  <.theme color_mode="dark" dark_theme="dark_dimmed">
+    Content
+  </.theme>
+  ```
+
+  Usually, the theme is set to user preferences. A menu with theme options can be created with `theme_menu_options/1` - see "Handling user selection" below.
+
   ## Persistency
-
-  There is no easy way to save persistent session data in LiveView because LiveView's state is stored in a process that ends when the page is left.
-
-  ### Session for theme state has been removed (as of 0.4.0)
-
-  Using the session to store the theme state (for example [using a AJAX call roundtrip](https://thepugautomatic.com/2020/05/persistent-session-data-in-phoenix-liveview/)) does not work as expected: when navigating to another LiveView page, the updated session data is not refetched and only becomes available after a page refresh. This means that the previous offered solution with `PrimerLive.ThemeSessionController` is no longer recommended and in fact removed.
-
-  ### Alternatives: database or cache
 
   If you already have a database set up for storing data by session ID, it's a small step to integrate the theme state with it.
 
@@ -25,6 +39,10 @@ defmodule PrimerLive.Theme do
 
   - [Cachex with Phoenix (2020)](https://www.alenm.com/code/phoenix-cachex)
   - [Use caching to speed up data loading in Phoenix LiveView](https://fullstackphoenix.com/quick_tips/liveview-caching)
+
+  ### Not recommended: Session for theme state
+
+  Using the session to store the theme state (for example [using a AJAX call roundtrip](https://thepugautomatic.com/2020/05/persistent-session-data-in-phoenix-liveview/)) does not work as expected: when navigating to another LiveView page, the updated session data is not refetched and only becomes available after a page refresh.
 
   ## Handling user selection
 
@@ -61,8 +79,6 @@ defmodule PrimerLive.Theme do
     {:noreply, socket}
   end
   ```
-
-
   """
 
   use Phoenix.Component